# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_kietle_zenreply_pkg_response.APIError:
    properties:
      code:
        type: string
      details:
        type: string
      message:
        type: string
    type: object
  github_com_kietle_zenreply_pkg_response.Meta:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  github_com_kietle_zenreply_pkg_response.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.APIError'
      message:
        type: string
      meta:
        $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Meta'
      success:
        type: boolean
    type: object
...
  handler.ListEntryRequest:
    properties:
      slack_user_id:
        example: U01234567
        type: string
    required:
    - slack_user_id
    type: object
...
  handler.UpdateSettingsRequest:
    properties:
      auto_reply_enabled:
        example: true
        type: boolean
      cooldown_minutes:
        example: 3
        type: integer
      default_message:
        example: I am in a deep work session, will reply soon.
        type: string
      default_reason:
        example: Deep Work
        type: string
      notify_on_resume:
        example: false
        type: boolean
      reply_in_thread:
        example: true
        type: boolean
    type: object
...
  model.UserSettings:
    properties:
      auto_reply_enabled:
        type: boolean
      blacklist:
        description: Blacklist is a JSON array of Slack user IDs that never receive
          auto-replies.
        items:
          type: string
        type: array
      cooldown_minutes:
        description: CooldownMinutes is the interval (in minutes) before re-sending
          an auto-reply to the same sender.
        type: integer
      created_at:
        type: string
      default_message:
        type: string
      default_reason:
        type: string
      id:
        type: string
      notify_on_resume:
        description: NotifyOnResume controls whether to send a follow-up when the
          user ends deep work.
        type: boolean
      reply_in_thread:
        description: ReplyInThread controls whether auto-replies are sent in the original
          message thread.
        type: boolean
      updated_at:
        type: string
      user_id:
        type: string
      whitelist:
        description: Whitelist is a JSON array of Slack user IDs that always receive
          auto-replies.
        items:
          type: string
        type: array
    type: object
host: localhost:8080
info:
  contact:
    email: support@zenreply.app
    name: ZenReply Support
    url: https://zenreply.app/support
  description: ZenReply is an intelligent Slack auto-reply system for deep work sessions.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://zenreply.app/terms
  title: ZenReply API
  version: 1.0.0
paths:
...
  /settings:
    get:
      description: Returns the authenticated user's auto-reply configuration
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.UserSettings'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Get user settings
      tags:
      - settings
    put:
      consumes:
      - application/json
      description: Updates the authenticated user's auto-reply configuration
      parameters:
      - description: Settings to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateSettingsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.UserSettings'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Update user settings
      tags:
      - settings
  /settings/blacklist:
    get:
      description: Returns the list of Slack user IDs that never receive auto-replies
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
            - properties:
                data:
                  items:
                    type: string
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Get blacklist
      tags:
      - settings
    post:
      consumes:
      - application/json
      description: Adds a Slack user ID to the blacklist. The user is also removed
        from the whitelist if present.
      parameters:
      - description: Slack user ID to add
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.ListEntryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Add to blacklist
      tags:
      - settings
  /settings/blacklist/{slack_user_id}:
    delete:
      description: Removes a Slack user ID from the blacklist
      parameters:
      - description: Slack User ID to remove
        in: path
        name: slack_user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Remove from blacklist
      tags:
      - settings
  /settings/reset:
    post:
      description: Resets all user settings to their default values
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.UserSettings'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Reset settings to defaults
      tags:
      - settings
  /settings/whitelist:
    get:
      description: Returns the list of Slack user IDs that always receive auto-replies
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
            - properties:
                data:
                  items:
                    type: string
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Get whitelist
      tags:
      - settings
    post:
      consumes:
      - application/json
      description: Adds a Slack user ID to the whitelist. The user is also removed
        from the blacklist if present.
      parameters:
      - description: Slack user ID to add
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.ListEntryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Add to whitelist
      tags:
      - settings
  /settings/whitelist/{slack_user_id}:
    delete:
      description: Removes a Slack user ID from the whitelist
      parameters:
      - description: Slack User ID to remove
        in: path
        name: slack_user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kietle_zenreply_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Remove from whitelist
      tags:
      - settings
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Enter: Bearer <your-jwt-token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"